+++ src/javascripts/models/user.js
@@ -1,13 +1,19 @@
-class User {
-  /**
-   * Constructor of User object
-   * @param {Object} data
-   */
-  constructor(data) {
-    this.id = data.id;
-    this.username = data.username;
-    this.password = data.password;
+import LoginService from "../services/loginService";
+
+class Users {
+  constructor() {
+    this.service = new LoginService();
+    this.userList;
   }
+
+  /**
+   * Method to get list of users from service
+   * @returns {Object} userList object
+   */
+  getUserList = async () => {
+    this.userList = await this.service.getUserList();
+    return this.userList;
+  };
 }
 
-export default User;
+export default Users;
